Create autosoundboard.py
I wrote this in 2014 to autogenerate soundboard apps for android. Just run this in the directory containing all your sound files with the names of the files the names that will be the buttons in the app. It generates the xml, manifest and java files to copy paste into android studio and compile. Instant sound board. I havent ran this in 10 years so no doubt it will break now. Still neat!